+++ StardewXnbHack/Program.cs
@@ -8,17 +8,18 @@ using StardewModdingAPI.Toolkit.Utilities;
 using StardewValley;
 using StardewXnbHack.Framework;
 using StardewXnbHack.Framework.Writers;
+using StardewXnbHack.ProgressHandling;
 
 namespace StardewXnbHack
 {
     /// <summary>The console app entry point.</summary>
-    internal class Program
+    public class Program
     {
         /*********
         ** Fields
         *********/
         /// <summary>The asset writers which support saving data to disk.</summary>
-        private readonly IAssetWriter[] AssetWriters = {
+        private static readonly IAssetWriter[] AssetWriters = {
             new MapWriter(),
             new SpriteFontWriter(),
             new TextureWriter(),
@@ -31,64 +32,84 @@ namespace StardewXnbHack
         ** Public methods
         *********/
         /// <summary>The console app entry method.</summary>
-        /// <param name="args">The arguments received by the app.</param>
-        static void Main(string[] args)
+        internal static void Main()
         {
-            new Program().Run();
+            Program.Run();
         }
 
         /// <summary>Unpack all assets in the content folder and store them in the output folder.</summary>
-        public void Run()
+        /// <param name="game">The game instance through which to unpack files, or <c>null</c> to launch a temporary internal instance.</param>
+        /// <param name="gamePath">The absolute path to the game folder, or <c>null</c> to auto-detect it.</param>
+        /// <param name="getLogger">Get a custom progress update logger, or <c>null</c> to use the default console logging. Receives the unpack context and default logger as arguments.</param>
+        public static void Run(Game1 game = null, string gamePath = null, Func<IUnpackContext, IProgressLogger, IProgressLogger> getLogger = null)
         {
+            // init logging
+            UnpackContext context = new UnpackContext();
+            IProgressLogger logger = new DefaultConsoleLogger(context);
+            if (getLogger != null)
+                logger = getLogger(context, logger);
+
             // get game info
             Platform platform = EnvironmentUtility.DetectPlatform();
-            string gamePath = new ModToolkit().GetGameFolders().FirstOrDefault()?.FullName;
-            if (gamePath == null)
+            context.GamePath = gamePath ?? new ModToolkit().GetGameFolders().FirstOrDefault()?.FullName;
+            if (context.GamePath == null || !Directory.Exists(context.GamePath))
             {
-                this.PrintColor("Can't find Stardew Valley folder.", ConsoleColor.Red);
+                logger.OnStartError("Can't find Stardew Valley folder.");
                 return;
             }
-            Console.WriteLine($"Found game folder: {gamePath}.");
-            Console.WriteLine();
 
             // get import/export paths
-            string contentPath = Path.Combine(gamePath, "Content");
-            string exportPath = Path.Combine(gamePath, "Content (unpacked)");
+            context.ContentPath = Path.Combine(context.GamePath, "Content");
+            context.ExportPath = Path.Combine(context.GamePath, "Content (unpacked)");
+            logger.OnStepChanged(ProgressStep.GameFound, $"Found game folder: {context.GamePath}.");
 
             // symlink files on Linux/Mac
             if (platform == Platform.Linux || platform == Platform.Mac)
             {
-                Process.Start("ln", $"-sf \"{Path.Combine(gamePath, "Content")}\"");
-                Process.Start("ln", $"-sf \"{Path.Combine(gamePath, "lib")}\"");
-                Process.Start("ln", $"-sf \"{Path.Combine(gamePath, "lib64")}\"");
+                foreach (string dirName in new[] { "Content", "lib", "lib64" })
+                {
+                    string fullPath = Path.Combine(context.GamePath, dirName);
+                    if (!Directory.Exists(fullPath))
+                        Process.Start("ln", $"-sf \"{fullPath}\"");
+                }
             }
 
-            // load game
-            ConsoleProgressBar progressBar;
-            Console.WriteLine("Loading game instance...");
-            Console.ForegroundColor = ConsoleColor.DarkGray;
-            using (Game1 game = this.GetGameInstance(platform, contentPath))
+            // load game instance
+            bool disposeGame = false;
+            if (game == null)
             {
-                Console.ResetColor();
-                Console.WriteLine();
-                Console.WriteLine("Unpacking files...");
+                logger.OnStepChanged(ProgressStep.LoadingGameInstance, "Loading game instance...");
+                game = Program.CreateTemporaryGameInstance(platform, context.ContentPath);
+                disposeGame = true;
+            }
+
+            // unpack files
+            try
+            {
+                logger.OnStepChanged(ProgressStep.UnpackingFiles, "Unpacking files...");
 
                 // collect files
-                DirectoryInfo contentDir = new DirectoryInfo(contentPath);
+                DirectoryInfo contentDir = new DirectoryInfo(context.ContentPath);
                 FileInfo[] files = contentDir.EnumerateFiles("*.xnb", SearchOption.AllDirectories).ToArray();
-                progressBar = new ConsoleProgressBar(files.Length);
+                context.Files = files;
 
                 // write assets
                 foreach (FileInfo file in files)
                 {
                     // prepare paths
-                    string assetName = file.FullName.Substring(contentPath.Length + 1, file.FullName.Length - contentPath.Length - 5); // remove root path + .xnb extension
-                    string fileExportPath = Path.Combine(exportPath, assetName);
+                    string assetName = file.FullName.Substring(context.ContentPath.Length + 1, file.FullName.Length - context.ContentPath.Length - 5); // remove root path + .xnb extension
+                    string relativePath = $"{assetName}.xnb";
+                    string fileExportPath = Path.Combine(context.ExportPath, assetName);
                     Directory.CreateDirectory(Path.GetDirectoryName(fileExportPath));
 
+                    // fallback logic
+                    void ExportRawXnb()
+                    {
+                        File.Copy(file.FullName, $"{fileExportPath}.xnb", overwrite: true);
+                    }
+
                     // show progress bar
-                    progressBar.Increment();
-                    progressBar.Print(assetName);
+                    logger.OnFileUnpacking(assetName);
 
                     // read asset
                     object asset = null;
@@ -98,8 +119,8 @@ namespace StardewXnbHack
                     }
                     catch (Exception ex)
                     {
-                        progressBar.Erase();
-                        this.PrintColor($"{assetName} => read error: {ex.Message}", ConsoleColor.Red);
+                        logger.OnFileUnpackFailed(relativePath, UnpackFailedReason.ReadError, $"read error: {ex.Message}");
+                        ExportRawXnb();
                         continue;
                     }
 
@@ -107,27 +128,24 @@ namespace StardewXnbHack
                     try
                     {
                         // get writer
-                        IAssetWriter writer = this.AssetWriters.FirstOrDefault(p => p.CanWrite(asset));
+                        IAssetWriter writer = Program.AssetWriters.FirstOrDefault(p => p.CanWrite(asset));
 
                         // write file
                         if (writer == null)
                         {
-                            progressBar.Erase();
-                            this.PrintColor($"{assetName}.xnb ({asset.GetType().Name}) isn't a supported asset type.", ConsoleColor.DarkYellow);
-                            File.Copy(file.FullName, $"{fileExportPath}.xnb", overwrite: true);
+                            logger.OnFileUnpackFailed(relativePath, UnpackFailedReason.UnsupportedFileType, $"{asset.GetType().Name} isn't a supported asset type.");
+                            ExportRawXnb();
                         }
-                        else if (!writer.TryWriteFile(asset, fileExportPath, assetName, platform, out string writeError))
+                        else if (!writer.TryWriteFile(asset, fileExportPath, assetName, platform,
+                            out string writeError))
                         {
-                            progressBar.Erase();
-                            this.PrintColor($"{assetName}.xnb ({asset.GetType().Name}) could not be saved: {writeError}.", ConsoleColor.DarkYellow);
-                            File.Copy(file.FullName, $"{fileExportPath}.xnb", overwrite: true);
+                            logger.OnFileUnpackFailed(relativePath, UnpackFailedReason.WriteError, $"{asset.GetType().Name} file could not be saved: {writeError}.");
+                            ExportRawXnb();
                         }
                     }
                     catch (Exception ex)
                     {
-                        Console.ForegroundColor = ConsoleColor.Red;
-                        Console.WriteLine($"{assetName} => export error: {ex.Message}");
-                        Console.ResetColor();
+                        logger.OnFileUnpackFailed(relativePath, UnpackFailedReason.UnknownError, $"unhandled export error: {ex.Message}");
                     }
                     finally
                     {
@@ -135,44 +153,50 @@ namespace StardewXnbHack
                     }
                 }
             }
+            finally
+            {
+                if (disposeGame)
+                    game.Dispose();
+            }
 
-            progressBar.Erase();
-            Console.WriteLine($"Done! Unpacked files to {exportPath}.");
+            logger.OnStepChanged(ProgressStep.Done, $"Done! Unpacked files to {context.ExportPath}.");
         }
 
 
         /*********
         ** Private methods
         *********/
-        /// <summary>Print a message to the console with a foreground color.</summary>
-        /// <param name="message">The message to print.</param>
-        /// <param name="color">The foreground color to use.</param>
-        private void PrintColor(string message, ConsoleColor color)
-        {
-            Console.ForegroundColor = color;
-            Console.WriteLine(message);
-            Console.ResetColor();
-        }
-
-        /// <summary>Get an initialised instance of Stardew Valley.</summary>
+        /// <summary>Create a temporary game instance for the unpacker.</summary>
         /// <param name="platform">The OS running the unpacker.</param>
         /// <param name="contentPath">The path to the content folder to import.</param>
-        private Game1 GetGameInstance(Platform platform, string contentPath)
+        private static Game1 CreateTemporaryGameInstance(Platform platform, string contentPath)
         {
-            Game1 game = new Game1();
+            var foregroundColor = Console.ForegroundColor;
+            Console.ForegroundColor = ConsoleColor.DarkGray;
 
-            if (platform == Platform.Windows)
+            try
             {
-                game.Content.RootDirectory = contentPath;
-                MethodInfo startGameLoop = typeof(Game1).GetMethod("StartGameLoop", BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.FlattenHierarchy);
-                if (startGameLoop == null)
-                    throw new InvalidOperationException("Can't locate game method 'StartGameLoop' to initialise internal game instance.");
-                startGameLoop.Invoke(game, new object[0]);
-            }
-            else
-                game.RunOneFrame();
 
-            return game;
+                Game1 game = new Game1();
+
+                if (platform == Platform.Windows)
+                {
+                    game.Content.RootDirectory = contentPath;
+                    MethodInfo startGameLoop = typeof(Game1).GetMethod("StartGameLoop", BindingFlags.Instance | BindingFlags.NonPublic | BindingFlags.FlattenHierarchy);
+                    if (startGameLoop == null)
+                        throw new InvalidOperationException("Can't locate game method 'StartGameLoop' to initialise internal game instance.");
+                    startGameLoop.Invoke(game, new object[0]);
+                }
+                else
+                    game.RunOneFrame();
+
+                return game;
+            }
+            finally
+            {
+                Console.ForegroundColor = foregroundColor;
+                Console.WriteLine();
+            }
         }
     }
 }